requirements.txt
langchain-openai
langchain_community
python-dotenv 
langchain_core
streamlit